Chatbot/requirements.txt
python-dotenv>=1.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
